% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{roxydocs-package}
\alias{roxydocs}
\alias{roxydocs-package}
\title{roxydocs: Documentation cpp11-generated wrapper functions}
\description{
Experimental package for documenting cpp11-generated wrapper functions using a new roxygen tag.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
